+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Unreleased
   type
   - Adjusted normalization logic to return references to cached build
     IDs if `cache_build_ids` is `true`
+- Added support for compressed debug information
+  - Added default enabled `zlib` feature
 - Introduced `normalize::Reason` enum to provide best guess at why normalization
   was not successful as part of the `normalize::UserMeta::Unknown` variant
 - Reduced number of allocations performed on address normalization and